fix(@n8n/performance): Use top-level await for engine init (beforeAll is broken in vitest bench)
vitest bench mode silently ignores beforeAll — errors are swallowed and
async setup never executes. Both engine blocks were running with the
default (current) engine, producing identical results.

Fix by splitting patterns.bench.ts into patterns-current.bench.ts and
patterns-vm.bench.ts, each using top-level await for engine initialization
with hard guard assertions. Extract shared benchmark definitions into
fixtures/pattern-benchmarks.ts to avoid duplication.

Also fix micro.bench.ts to use top-level await instead of beforeAll.
